backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
uvloop>=0.14.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.23
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
alembic>=1.7.4
pytest>=6.2.5
httpx>=0.23.0
openai>=0.27.0
mutagen>=1.45.1
openai-whisper>=20231117
google-generativeai>=0.3.0 